Remove resources added to ocp_python_wrapper
